Implement custom debug for use with Termination
The current implementation of the Termination trait results in very
gross output if used. There seem to be two options for this:

1. Have a `run` function which contains the real logic of the program.
2. Have the debug output for the `Error` struct be bad for anything
   except the Termination trait.

As this library is targeting humans and the error probably shouldn't be
held by any other object (it should be immediately bubbled up), this
commit implements the second option.