Use `Cow` instead of `String`
I suspect it will be far more common to use static strings over
formatted strings when using this library, so it might be worth not
allocating when using a static string.

But that might not really matter in the use-case for this library, so I
might sit on it for a while.