Add caller information to the error
This is not meant for production - it is an easy way to have some kind
of error reporting during hacking or for developer/internal tools. As
such, having the line that reported the error visible on the output will
be useful.

Unfortunately, it isn't available until Rustc 1.46 meaning this might
not get push to Crates.io for a hot second. Or maybe it will - maybe I
should change my policies from "this crate will support the oldest
Ubuntu LTS" to "changes will be backported." That would be useful to me
but likely confusing to the "community."