Prepare v1.0 release
I don't know if I really consider this crate "done" but experience has
shown me that it does what I want and I am unlikely to change it at this
point. So far, all of my Rust work hasn't needed any other error
handling and I doubt that will change in the near future.

...Plus, there are other crates to fill this area now and I don't see a
need to try and compete with them. This is my own little niche and I'll
stick to it.